Make hashtag optional in hex input (#3)
Verifies both the original hex string and a modified one to see if it's valid. This allows for input with and without the hashtag, while still allowing you to write color names.

Also fixes the issue where the input is only partially parsed because the partial input is already valid. This is done by only setting the color (if valid) on OnSubmit and OnDeselect. This also makes it mimic the default behavior of color text input fields